geotiff: extract read_geotiff_gpu to _backends/gpu.py (#1885)
Step 7 of #1813's multi-PR refactor of __init__.py. Pure code motion;
no public API change.

Moved into a new xrspatial/geotiff/_backends/gpu.py:

- read_geotiff_gpu: GPU read entry point (~640 lines, the largest
  single function in __init__.py).
- _gds_chunk_path_available: predicate gating the GDS direct-to-GPU
  chunked path.
- _decode_window_gpu_direct: per-chunk disk->GPU decode used by the
  GDS path.
- _read_geotiff_gpu_chunked: lazy Dask+CuPy graph builder for
  read_geotiff_gpu(chunks=...).
- _read_geotiff_gpu_chunked_gds: direct-to-GPU specialisation built
  by _read_geotiff_gpu_chunked when KvikIO + a chunky tiled local
  file qualifies.

Function bodies are unchanged except for adjusted relative imports
(``._X`` -> ``.._X`` outside _backends, ``._gpu_helpers`` for the
already-extracted GPU helpers). __init__.py re-exports
read_geotiff_gpu so ``from xrspatial.geotiff import read_geotiff_gpu``
keeps working. The four private helpers are not re-exported because
they were only called by read_geotiff_gpu itself; nothing outside
the GPU module references them.

One known consumer of the private names had to update its monkeypatch
target: test_gpu_chunks_out_of_core_1876.py patches
``_decode_window_gpu_direct`` to spy on chunk-decode calls. The patch
target moves from ``xrspatial.geotiff`` to
``xrspatial.geotiff._backends.gpu`` so the spy still intercepts the
real call site (module-global lookup happens in the defining module).

The Dask+CuPy chunked-fallback path needs ``read_geotiff_dask``,
which still lives in __init__.py. To avoid a circular import,
_read_geotiff_gpu_chunked lazy-imports it inside the function body
(``from .. import read_geotiff_dask``). PR #8 will move
``read_geotiff_dask`` into ``_backends/dask.py`` and the lazy import
can be replaced with a static one then.

Net __init__.py change: 3969 -> 2962 lines (-1007). _backends/gpu.py
is 1047 lines.

Verification: pixel-parity matrix from #1889 (192 cells, all GPU
cells), runtime identity (9), GPU byteswap (#1508), GPU kwarg rename
(#1560), GPU chunks out-of-core (#1876, including the GDS spy
tests), full geotiff suite minus 8 pre-existing main failures
(read_to_array monkeypatch + tile_size_positive_works) -- 2862/2862
pass.

Refs #1813.
